--- src/main/java/InvalidTodo.java
@@ -1,5 +1,0 @@
-public class InvalidTodo extends DukeExceptions {
-    public InvalidTodo() {
-        super("ToDo invalid. Please fill in your todo properly eg. todo [description]");
-    }
-}
