+++ src/main/java/duke/exceptions/DukeExceptions.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class DukeExceptions extends Exception{
     public DukeExceptions(String msg) {
         super(msg);
