+++ src/main/java/duke/exceptions/InvalidDeadline.java
@@ -1,6 +1,8 @@
+package duke.exceptions;
+
 public class InvalidDeadline extends DukeExceptions {
     public InvalidDeadline() {
-        super("Deadline invalid. Please fill in your deadline properly eg. deadline [description] /by [by]");
+        super("duke.tasks.Deadline invalid. Please fill in your deadline properly eg. deadline [description] /by [by]");
     }
 
     public InvalidDeadline(String msg) {
