+++ src/main/java/duke/exceptions/InvalidEvent.java
@@ -1,6 +1,8 @@
+package duke.exceptions;
+
 public class InvalidEvent extends DukeExceptions {
     public InvalidEvent() {
-        super("Event invalid. Please fill in your event properly eg. event [description] /from [from] /to [to]");
+        super("duke.tasks.Event invalid. Please fill in your event properly eg. event [description] /from [from] /to [to]");
     }
 
     public InvalidEvent(String msg) {
