+++ src/main/java/duke/exceptions/InvalidInput.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class InvalidInput extends DukeExceptions {
     public InvalidInput() {
         super("Sorry, I did not understand your instructions, please try again.");
