+++ text-ui-test/EXPECTED.TXT
@@ -7,9 +7,9 @@ Hello from
 
 Greetings
 added: aaa (from: 2 to: 3)
-Now you have 1 tasks in the list
+Now you have 1 duke.tasks in the list
 added: bbb (by: hello!)
-Now you have 2 tasks in the list
+Now you have 2 duke.tasks in the list
 1. [E][ ] aaa (from: 2 to: 3)
 2. [D][ ] bbb (by: hello!)
 Marked:
